Move batch db code under batchload
